Add admin role check for retrieving case files
This commit adds an admin role check to the function responsible for retrieving case files in the SingleCasePage component. If the user has an admin role, the function will call the getAdminCaseFilesById method; otherwise, it will call the getCaseFilesById method. This change ensures that different case files are retrieved based on the user's role.